api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
exifread
sqlalchemy
psycopg2-binary
pydantic
celery
redis
onnxruntime
numpy

--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
opencv-python-headless
matplotlib
timm
transformers==4.56.2
addict
pyyaml
yapf